Factor out extraction helpers module
